+++ Editor/Aseprite/AsepriteUtilities.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.IO;
 using Unity.Collections;
+using Unity.Collections.LowLevel.Unsafe;
 using UnityEngine;
 
 namespace UnityEditor.U2D.Aseprite
@@ -46,25 +47,35 @@ namespace UnityEditor.U2D.Aseprite
             if (colorDepth == 32)
             {
                 image = new NativeArray<Color32>(data.Length / 4, Allocator.Persistent);
-                for (var i = 0; i < image.Length; ++i)
+                var imageLength = image.Length;
+                unsafe
                 {
-                    var dataIndex = i * 4;
-                    image[i] = new Color32(
-                        data[dataIndex],
-                        data[dataIndex + 1],
-                        data[dataIndex + 2],
-                        data[dataIndex + 3]);
+                    var imageArr = (Color32*)image.GetUnsafePtr();
+                    for (var i = 0; i < imageLength; ++i)
+                    {
+                        var dataIndex = i * 4;
+                        imageArr[i] = new Color32(
+                            data[dataIndex],
+                            data[dataIndex + 1],
+                            data[dataIndex + 2],
+                            data[dataIndex + 3]);
+                    }
                 }
             }
             else if (colorDepth == 16)
             {
                 image = new NativeArray<Color32>(data.Length / 2, Allocator.Persistent);
-                for (var i = 0; i < image.Length; ++i)
+                var imageLength = image.Length;
+                unsafe
                 {
-                    var dataIndex = i * 2;
-                    var value = data[dataIndex];
-                    var alpha = data[dataIndex + 1];
-                    image[i] = new Color32(value, value, value, alpha);
+                    var imageArr = (Color32*)image.GetUnsafePtr();
+                    for (var i = 0; i < imageLength; ++i)
+                    {
+                        var dataIndex = i * 2;
+                        var value = data[dataIndex];
+                        var alpha = data[dataIndex + 1];
+                        imageArr[i] = new Color32(value, value, value, alpha);
+                    }
                 }
             }
             return image;